Fix for #26385
The `assign` annotation can leave a dangling pointer, which is why Xcode wants the addition of `__unsafe_unretained` on the instance variable in addition. In this case (delegates), it’s best to use `weak`, which will avoid retain cycles _and_ set the pointer to `nil` if the object is released.